Extract Multipart handler methods to MultipartController
