+++ server/src/test/java/com/adobe/testing/s3mock/FileStoreControllerTest.java
@@ -17,17 +17,11 @@
 package com.adobe.testing.s3mock;
 
 import static com.adobe.testing.s3mock.S3Exception.BAD_REQUEST_MD5;
-import static com.adobe.testing.s3mock.S3Exception.ENTITY_TOO_SMALL;
-import static com.adobe.testing.s3mock.S3Exception.INVALID_PART;
-import static com.adobe.testing.s3mock.S3Exception.INVALID_PART_ORDER;
-import static com.adobe.testing.s3mock.S3Exception.NO_SUCH_UPLOAD_MULTIPART;
 import static com.adobe.testing.s3mock.util.AwsHttpHeaders.CONTENT_MD5;
 import static com.adobe.testing.s3mock.util.AwsHttpHeaders.X_AMZ_SERVER_SIDE_ENCRYPTION;
 import static com.adobe.testing.s3mock.util.AwsHttpHeaders.X_AMZ_SERVER_SIDE_ENCRYPTION_AWS_KMS_KEY_ID;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.ArgumentMatchers.anyMap;
-import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.contains;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.ArgumentMatchers.isNull;
@@ -35,28 +29,19 @@ import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.head;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
 
 import com.adobe.testing.s3mock.dto.Bucket;
-import com.adobe.testing.s3mock.dto.CompleteMultipartUpload;
-import com.adobe.testing.s3mock.dto.ErrorResponse;
-import com.adobe.testing.s3mock.dto.Part;
 import com.adobe.testing.s3mock.service.BucketService;
 import com.adobe.testing.s3mock.service.ObjectService;
 import com.adobe.testing.s3mock.store.BucketStore;
 import com.adobe.testing.s3mock.store.KmsKeyStore;
 import com.adobe.testing.s3mock.store.S3ObjectMetadata;
 import com.adobe.testing.s3mock.util.DigestUtil;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.dataformat.xml.XmlMapper;
 import java.io.File;
 import java.io.InputStream;
 import java.nio.file.Paths;
 import java.time.Instant;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
 import org.apache.commons.io.FileUtils;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -74,8 +59,6 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 @MockBeans({@MockBean(classes = {KmsKeyStore.class, BucketStore.class})})
 @SpringBootTest(classes = {S3MockConfiguration.class})
 class FileStoreControllerTest {
-  private static final ObjectMapper MAPPER = new XmlMapper();
-
   private static final String TEST_BUCKET_NAME = "test-bucket";
   private static final Bucket TEST_BUCKET =
       new Bucket(Paths.get("/tmp/foo/1"), TEST_BUCKET_NAME, Instant.now().toString());
@@ -156,121 +139,6 @@ class FileStoreControllerTest {
   }
 
   @Test
-  void testCompleteMultipart_BadRequest_uploadTooSmall() throws Exception {
-    givenBucket();
-    List<Part> parts = new ArrayList<>();
-    parts.add(createPart(0, 5L));
-    parts.add(createPart(1, 5L));
-
-    CompleteMultipartUpload uploadRequest = new CompleteMultipartUpload();
-    for (Part part : parts) {
-      uploadRequest.setPart(part);
-    }
-
-    String key = "sampleFile.txt";
-    String uploadId = "testUploadId";
-    doThrow(ENTITY_TOO_SMALL)
-        .when(objectService)
-        .verifyMultipartParts(eq(TEST_BUCKET_NAME), eq(key), eq(uploadId), anyList());
-    ErrorResponse errorResponse = from(ENTITY_TOO_SMALL);
-
-    mockMvc.perform(
-            post("/test-bucket/" + key)
-                .accept(MediaType.APPLICATION_XML)
-                .content(MAPPER.writeValueAsString(uploadRequest))
-                .param("uploadId", uploadId)
-        ).andExpect(MockMvcResultMatchers.status().isBadRequest())
-        .andExpect(MockMvcResultMatchers.content().xml(MAPPER.writeValueAsString(errorResponse)));
-  }
-
-  @Test
-  void testCompleteMultipart_BadRequest_uploadIdNotFound() throws Exception {
-    givenBucket();
-    String uploadId = "testUploadId";
-
-    List<Part> parts = new ArrayList<>();
-    parts.add(createPart(0, 5L));
-    parts.add(createPart(1, 5L));
-
-    doThrow(NO_SUCH_UPLOAD_MULTIPART)
-        .when(objectService)
-        .verifyMultipartParts(eq(TEST_BUCKET_NAME), anyString(), eq(uploadId), anyList());
-
-    CompleteMultipartUpload uploadRequest = new CompleteMultipartUpload();
-    for (Part part : parts) {
-      uploadRequest.setPart(part);
-    }
-    ErrorResponse errorResponse = from(NO_SUCH_UPLOAD_MULTIPART);
-
-    String key = "sampleFile.txt";
-    mockMvc.perform(
-            post("/test-bucket/" + key)
-                .accept(MediaType.APPLICATION_XML)
-                .content(MAPPER.writeValueAsString(uploadRequest))
-                .param("uploadId", uploadId)
-        ).andExpect(MockMvcResultMatchers.status().isNotFound())
-        .andExpect(MockMvcResultMatchers.content().xml(MAPPER.writeValueAsString(errorResponse)));
-  }
-
-  @Test
-  void testCompleteMultipart_BadRequest_partNotFound() throws Exception {
-    givenBucket();
-    String key = "sampleFile.txt";
-    String uploadId = "testUploadId";
-
-    List<Part> requestParts = new ArrayList<>();
-    requestParts.add(createPart(1, 5L));
-
-    doThrow(INVALID_PART)
-        .when(objectService)
-        .verifyMultipartParts(eq(TEST_BUCKET_NAME), eq(key), eq(uploadId), anyList());
-
-    CompleteMultipartUpload uploadRequest = new CompleteMultipartUpload();
-    for (Part part : requestParts) {
-      uploadRequest.setPart(part);
-    }
-    ErrorResponse errorResponse = from(INVALID_PART);
-
-    mockMvc.perform(
-            post("/test-bucket/" + key)
-                .accept(MediaType.APPLICATION_XML)
-                .content(MAPPER.writeValueAsString(uploadRequest))
-                .param("uploadId", uploadId)
-        ).andExpect(MockMvcResultMatchers.status().isBadRequest())
-        .andExpect(MockMvcResultMatchers.content().xml(MAPPER.writeValueAsString(errorResponse)));
-  }
-
-  @Test
-  void testCompleteMultipart_BadRequest_invalidPartOrder() throws Exception {
-    givenBucket();
-
-    String key = "sampleFile.txt";
-    String uploadId = "testUploadId";
-
-    doThrow(INVALID_PART_ORDER)
-        .when(objectService)
-        .verifyMultipartParts(eq(TEST_BUCKET_NAME), eq(key), eq(uploadId), anyList());
-
-    List<Part> requestParts = new ArrayList<>();
-    requestParts.add(createPart(1, 5L));
-    requestParts.add(createPart(0, 5L));
-
-    CompleteMultipartUpload uploadRequest = new CompleteMultipartUpload();
-    for (Part part : requestParts) {
-      uploadRequest.setPart(part);
-    }
-    ErrorResponse errorResponse = from(INVALID_PART_ORDER);
-
-    mockMvc.perform(
-            post("/test-bucket/" + key)
-                .accept(MediaType.APPLICATION_XML)
-                .content(MAPPER.writeValueAsString(uploadRequest))
-                .param("uploadId", uploadId)
-        ).andExpect(MockMvcResultMatchers.status().isBadRequest())
-        .andExpect(MockMvcResultMatchers.content().xml(MAPPER.writeValueAsString(errorResponse)));
-  }
-
-  @Test
   void testGetObject_Encrypted_Ok() throws Exception {
     String encryption = "aws:kms";
     String encryptionKey = "key-ref";
@@ -319,15 +187,6 @@ class FileStoreControllerTest {
     ).andExpect(MockMvcResultMatchers.status().isNotFound());
   }
 
-  private Part createPart(int partNumber, long size) {
-    Part part = new Part();
-    part.setPartNumber(partNumber);
-    part.setSize(size);
-    part.setLastModified(new Date());
-    part.setETag("someEtag" + partNumber);
-    return part;
-  }
-
   private void givenBucket() {
     when(bucketService.getBucket(TEST_BUCKET_NAME)).thenReturn(TEST_BUCKET);
     when(bucketService.doesBucketExist(TEST_BUCKET_NAME)).thenReturn(true);
@@ -352,13 +211,6 @@ class FileStoreControllerTest {
     final File sourceFile = new File("src/test/resources/sampleFile.txt");
     s3ObjectMetadata.setDataPath(sourceFile.toPath());
     return s3ObjectMetadata;
-  }
-
-  private ErrorResponse from(S3Exception e) {
-    ErrorResponse errorResponse = new ErrorResponse();
-    errorResponse.setCode(e.getCode());
-    errorResponse.setMessage(e.getMessage());
-    return errorResponse;
   }
 }
